COMP: Fix sign comparison warnings
To identify lines triggering a compilation warning, the
following one-liner was used:

  ack-grep "for\s*\(\s+int\s+[a-z]+\s*\=.+ImageDimension"

Also updates ITKMinimalPathExtraction:

$ git shortlog aed93f4..c9cdc87 --no-merges
Jean-Christophe Fillion-Robin (1):
      Fix sign comparison warnings in SingleImageCostFunction

Fix warnnnnnn signed

+++ Applications/ConvertTubesToDensityImage/itktubeTubeSpatialObjectToDensityImage.hxx
@@ -33,7 +33,7 @@ TubeSpatialObjectToDensityImage< TDensityImageType, TRadiusImageType,
                                  TTangentImageType >
 ::TubeSpatialObjectToDensityImage( void )
 {
-  for(short i = 0; i < ImageDimension; i++ )
+  for(unsigned int i = 0; i < ImageDimension; i++ )
     {
     m_Size[i] = 0;
     m_Spacing[i] = 1;
@@ -142,7 +142,7 @@ TubeSpatialObjectToDensityImage< TDensityImageType, TRadiusImageType,
       {
       VectorPixelType v = it_vector.Value();
       typename DensityImageType::IndexType index = it_vector.GetIndex();
-      for( int i = 0; i < ImageDimension; i++ )
+      for( unsigned int i = 0; i < ImageDimension; i++ )
         {
         index[i] += v[i];
         }
@@ -167,7 +167,7 @@ TubeSpatialObjectToDensityImage< TDensityImageType, TRadiusImageType,
       {
       VectorPixelType v = it_vector.Value();
       typename DensityImageType::IndexType index = it_vector.GetIndex();
-      for( int i = 0; i < ImageDimension; i++ )
+      for( unsigned int i = 0; i < ImageDimension; i++ )
         {
         index[i] += v[i];
         }
